+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/ContentStateReducer.kt
@@ -198,6 +198,9 @@ internal object ContentStateReducer {
             is ContentAction.SetRecordingDevices -> updateContentState(state, action.sessionId) {
                 it.copy(recordingDevices = action.devices)
             }
+            is ContentAction.UpdateDesktopModeAction -> updateContentState(state, action.sessionId) {
+                it.copy(desktopMode = action.enabled)
+            }
         }
     }
 }